Add basic Product class
This will be the basic unit that gets loaded into the vending machine and vended out. It has a name and a price that will be used when the user requests a product.

I debated whether to make this into a ‘collection’ of products, i.e. give it a quantity. However I felt like this was a bit less representative of an actual product, that’s more of a ‘product row’ than a product. It also adds undue complexity at this point. I may add that in in the future if it presents itself as a better solution.