Add system for inserting coins into vending machine
This’ll take in coins and store them in an array. There’s no way to see which coins you’ve inserted (like a real vending machine) but you can see your balance which just returns a sum of the inserted coins value. To the outside they all go into one value but internally we are tracking the coins returned so they can be added to the change pile for future purchases, returned during the current transaction, or returned to the user if they cancel their transaction (if we get time to implementing that).