Add the ability to insert coin in CLI
I initially built this so that the the user_input case statement in CLI.run had a number of separate options for each different coin, so it was similar to `when ‘insert £1’, ‘insert £2’` (though dynamically generated).

The user input then got passed through into insert_coin where it was split apart and then the option was used as it currently is.

This felt a little off however and so I tried out swapping over to splitting the command and option apart prior to triggering the case statement. This lead to a much nicer system that will provide a smoother implementation of the purchasing flow later down the line.