Set up specs for the CLI
To test the CLI these specs simulate sending input by stubbing calls to Kernel.gets with a series of strings.

This will cause the system to immediately loop through all the input, after which we can capture the output and compare it to what we expect to have happened. Since the system is deterministic we can be sure of what we expect to get back.