Fix bug where exit does not ignore options
We had swapped over to using command (the first word in the user_input) for determing what action to perform. Unfortunately this wasn’t being used for determining whether to close the program, it was still using the full user_input. This meant if you entered something like `exit foo` it would not close the program.