Add initial stock/change loading
This adds in a config file that handles the storage of the initial stock/change in the vending machine. This is parsed by ConfigLoader which is used within the CLI for loading the VendingMachine.

Abstracting the config loading out into a class of its own has allowed me to add some basic validation during loading to ensure that we get proper data passed through.

A part of this was realisign that having Coin.from_string return an invalid_coin when the string was not a valid coin name lead to having invalid coins floating around the system. By swapping it over to nil and instead compacting what gets passed into the vending machine during set up we can be sure no invalid coins are created and used. We should avoid creating invalid coins completely.

To allow the specs to pick up on the loading this also adds in the stock command that outputs the current stock.