Add ability to purchase products
This adds in the request_product function to the VendingMachine.

Rather than returning just a product or false this responds a VendingResponse object that includes all the information required to discern what has happened. It will include info on the change returned to the user along with the vended product in cases where the purchase was successful. Where there is an issue this will have no change or vended product but will respond to .error to explain what went wrong.