Add info to README about Time
Info on Time Taken and what I’d do with extra time.

Email described these options for how much time to taken:
```
We appreciate some people have commitments and have varying availability in terms of the amount of time they can commit to the technical exercise so there would be two options.

Option 1
Time boxed for 2-3 hours. If you pick this please, make sure it's clearly stated that this is the option you picked, how much time you took and what you would have done given more time. Make sure it is clear in the Readme.

Option 2
If you want to spend more time on the excercise ie. 4-5 hours, make sure this is the option you picked is clearly detailed in the Readme and include how much time you spend on it.
```