Remove references to @vars in vending_machine.rb
I ran into an issue during the initial implementation of this where I was trying to run `change -= removed_change`. This was throwing this error:
 ```
NoMethodError:
        undefined method `-' for nil:NilClass
```

This was taking a while to resolve so I flipped to just using `@change -= removed_change`. The same issue presented itself in the product removal.
However, as I was adding in more specs I realised that this code actually caused a bug where it’d removed multiple instances of an coin or product rather than just a single instance. To fix that I moved to using delete_at with a specific index. This fixed the multi-deletion issue however I didn’t realise at the time that it allowed me to move back to interacting with product and change rather than messily going directly to the @ variables.

I realised this as I was writing documentation and flipped it back, the specs still work and a sanity test via the CLI confirmed everything is still working.